init fix
