+++ config/config-models.go
@@ -1,0 +1,22 @@
+package config
+
+type RequestService struct {
+	DatabaseUrl     string `yaml:"database_url"`
+	DatabaseType    string `yaml:"database_type"`
+	RestServiceHost string `yaml:"rest_service_host"`
+	RestServicePort string `yaml:"rest_service_port"`
+}
+
+type OAuthProviderConfig struct {
+	ClientID     string   `yaml:"client_id"`
+	ClientSecret string   `yaml:"client_secret"`
+	UserInfoUrl  string   `yaml:"user_info_url"`
+	RedirectUrl  string   `yaml:"redirect_url"`
+	Scopes       []string `yaml:"scopes"`
+}
+
+type UserService struct {
+	AllowedHosts     []string            `yaml:"allowed_hosts"`
+	GoogleProvider   OAuthProviderConfig `yaml:"google_provider"`
+	FacebookProvider OAuthProviderConfig `yaml:"facebook_provider"`
+}
